Switch to float64

+++ src/environments/PlantGrowthChamber/utils.py
@@ -32,7 +32,7 @@ async def get_session():
 
 
 def get_one_hot_time_observation(local_time: datetime):
-    one_hot = np.zeros(13, dtype=np.float32)
+    one_hot = np.zeros(13, dtype=np.float64)
     index = max(0, min(12, local_time.hour - 9))
     one_hot[index] = 1.0
     return one_hot
